% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{add_ptt_attr}
\alias{add_ptt_attr}
\title{Re-add ptt attributes back}
\usage{
add_ptt_attr(to, from)
}
\arguments{
\item{to}{A df to put attributes}

\item{from}{A df to get attributes froms}
}
\value{

}
\description{
Re-add ptt attributes back
}
