% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{agg_key}
\alias{agg_key}
\title{Aggregate based on key}
\usage{
agg_key(
  x,
  by = NULL,
  .fns = sum,
  value_cols = c("values", "value"),
  w = NULL,
  na.rm = FALSE,
  all_to_regions = TRUE,
  key = NULL
)
}
\arguments{
\item{x}{A data.frame like object.}

\item{by}{A character vector of variables to join by.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{key}{A data.frame. A classification key in same form as
as key from \code{\link[statficlassifications]{get_regionkey}}.}
}
\description{
Aggregate based on key
}
\examples{
x <- data.frame(alue_code = c("SSS", "KU049", "KU091", "KU109"), values = c(1,1,1,2))
agg_regions(x, na.rm = TRUE)
agg_regions(x, na.rm = TRUE, pass_region_codes = "SSS")
agg_regions(x, na.rm = TRUE, pass_region_codes = "SSS", all_to_regions = FALSE)
z <- dplyr::mutate(x, values2 = c(3,4,5,6))
agg_regions(z, na.rm = TRUE, value_cols = c("values", "values2"))
agg_key(x, na.rm = TRUE,
            key = data.frame(
               alue_code = c("SSS", "KU049", "KU091", "KU109"),
               maakunta_name = c("SSS", "Maakunta1", "Maakunta1", "Maakunta2"))
               )
agg_key(x,
            .fns = mean,
            key = data.frame(
               alue_code = c("SSS", "KU049", "KU091", "KU109"),
               maakunta_name = c("SSS", "Maakunta1", "Maakunta1", "Maakunta2"))
               )
xx <- data.frame(alue_code = c("SSS", "KU049", "KU091", "KU109"), values = c(1,2,1,1), size = c(1,2,1,1))
agg_key(xx,
            .fns = weighted.mean,
            w = size,
            key = data.frame(
               alue_code = c("SSS", "KU049", "KU091", "KU109"),
               maakunta_name = c("SSS", "Maakunta1", "Maakunta1", "Maakunta2"))
               )
}
