% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{agg_regions}
\alias{agg_regions}
\alias{add_regional_agg}
\title{Aggregate to other region level}
\usage{
agg_regions(
  x,
  from = "kunta",
  to = "maakunta",
  value_cols = c("values"),
  pass_region_codes = NULL,
  na.rm = FALSE,
  all_to_regions = TRUE,
  custom_key = NULL,
  by_name = FALSE
)

add_regional_agg(x, from = "kunta", to = "maakunta")
}
\arguments{
\item{x}{A data.frame like object.}

\item{from}{A name of reginal classification.}

\item{to}{A name of reginal classification.}

\item{value_cols}{A character vector of name of columns to aggregate.}

\item{pass_region_codes}{Region codes to be passed true. For example whole
country "SSS". Does not affect aggregation. Default NULL.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{all_to_regions}{logical. Should all to regions included
even if not in data.}

\item{custom_key}{A data.frame. Custom classification key in same form as
as key from \code{\link[statficlassifications]{get_regionkey}}.}

\item{by_name}{A locigal to join by name or code (default).}
}
\description{
Aggregate to other region level

Add regional aggragation
}
\section{Functions}{
\itemize{
\item \code{add_regional_agg()}: 

}}
\examples{
x <- data.frame(alue_code = c("SSS", "KU049", "KU091", "KU109"),
                alue_name = c("KOKO MAA", "Espoo", "Helsinki", "Hämeenlinna"), values = c(1,1,1,2))
agg_regions(x, na.rm = TRUE)
agg_regions(x, na.rm = TRUE, pass_region_codes = "SSS")
agg_regions(x, na.rm = TRUE, pass_region_codes = "SSS", all_to_regions = FALSE)
z <- dplyr::mutate(x, values2 = c(3,4,5,6))
agg_regions(z, na.rm = TRUE, value_cols = c("values", "values2"))
agg_regions(x, na.rm = TRUE,
            custom_key = data.frame(
               kunta_code = c("SSS", "KU049", "KU091", "KU109"),
               kunta_name = c("KOKO MAA", "Espoo", "Helsinki", "Hämeenlinna"),
               maakunta_code = as.factor(c("SSS", "MK1", "MK1", "MK2")),
               maakunta_name = as.factor(c("SSS", "Maakunta1", "Maakunta1", "Maakunta2")))
               )
agg_regions(x, na.rm = TRUE,
            custom_key = data.frame(
               kunta_code = c("SSS", "KU049", "KU091", "KU109"),
               kunta_name = c("KOKO MAA", "Espoo", "Helsinki", "Hämeenlinna"),
               maakunta_code = as.factor(c("SSS", "MK1", "MK1", "MK2")),
               maakunta_name = as.factor(c("SSS", "Maakunta1", "Maakunta1", "Maakunta2"))),
               by_name = TRUE
               )
}
