% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extra.R
\name{alevels}
\alias{alevels}
\title{List of levels of all factors in object}
\usage{
alevels(x)
}
\arguments{
\item{x}{an object}
}
\value{
Namede list of factory levels.
}
\description{
Work at least for data.frames
}
\keyword{utilities}
