% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extra.R
\name{conc}
\alias{conc}
\title{To copy to the clipboard aka Ctrl + c}
\usage{
conc(x)
}
\arguments{
\item{x}{An object}
}
\description{
To copy an object to the clipboard.
}
\details{
To mimic ctrl + c. Currently defined as \code{conc <- function(x)
write.table(x, "clipboard-128", sep="\t", dec=",", col.names=NA, row.names =
if (is.ts(x)) gsub('\.',",",as.character(time(x))) else TRUE)}
}
\examples{
  \dontrun{
   x <- data.frame(a = c(1,2), b = c(3,4))
   conc(x)
  }
}
\seealso{
\code{\link{write.table}}
}
\keyword{manip}
\keyword{utilities}
