% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extra.R
\name{factor_all}
\alias{factor_all}
\title{Mutate all character variables to factor with as_factor().}
\usage{
factor_all(.data)
}
\arguments{
\item{.data}{}
}
\description{
Mutate all character variables to factor with as_factor().
}
\examples{
df <- data.frame(a = c("a", "b"), b = c(1,2))
df <- factor_all(df)
str(df)

}
