% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_save_load.R
\name{filter_region_level}
\alias{filter_region_level}
\title{Filter region level}
\usage{
filter_region_level(data, region_level)
}
\arguments{
\item{data}{data.frame from \code{ptt_read_data}-function with \code{alue_code}-column.}

\item{region_level}{\code{"kunta"}, \code{"seutukunta"}, \code{"maakunta"} or \code{"suuralue"}.
The desired region level}
}
\value{

}
\description{
Filters, from a data.frame with a reigon code variable (with standardized region code prefixes)
desired region level and renames the the region variables in case of unique region level.
}
