% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_names.R
\name{get_codes_names}
\alias{get_codes_names}
\title{Get codes-names key}
\usage{
get_codes_names(x, db_list_name = NULL, variable = NULL, offline = TRUE)
}
\arguments{
\item{x}{chr, table_code or url. If table code function requires db_list_name}

\item{db_list_name}{chr, db_list_name}

\item{variable}{chr, (statfitools::make_names-cleaned) variable name}

\item{offline}{logical, whether to try look for pxweb_metadata in the db_list_offline}
}
\value{
A named (column codes) list of named (codes) vectors (names).
}
\description{
Note that a list is only sensible form of having code-name keys for different
variables as different variables have a different number of these keys. Thus,
this function produces lists of named vectors for many variables and named
vectors for one variable.
}
