% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_search.R
\name{get_pxweb_metadata}
\alias{get_pxweb_metadata}
\alias{get_variables}
\alias{get_time_var}
\alias{get_manual_metadata}
\alias{get_title}
\title{Get information about data in db_list}
\usage{
get_pxweb_metadata(table_code, db_list_name)

get_variables(table_code, db_list_name, as_string = FALSE)

get_time_var(table_code, db_list_name)

get_manual_metadata(table_code, db_list_name)

get_title(table_code, db_list_name)
}
\arguments{
\item{table_code}{chr}

\item{db_list_name}{chr}
}
\description{
Get information about data in db_list
}
\section{Functions}{
\itemize{
\item \code{get_variables()}: Get information about data in db_list

\item \code{get_time_var()}: Get information about data in db_list

\item \code{get_manual_metadata()}: Get information about data in db_list

\item \code{get_title()}: Get information about data in db_list

}}
