% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statfi.R
\name{is_table_code}
\alias{is_table_code}
\title{Test if table code in existing db_list}
\usage{
is_table_code(x)
}
\arguments{
\item{x}{chr, potential table code}
}
\value{
logical
}
\description{
Test if table code in existing db_list
}
