% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extra.R
\name{pc}
\alias{pc}
\title{Calculate percentage change}
\usage{
pc(x, n, order_by = time)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Positive integer of length 1, giving the number of positions to lead or lag by}

\item{order_by}{Override the default ordering to use another vector or column}
}
\description{
Calculate percentage change
}
