% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{print_full_filter}
\alias{print_full_filter}
\alias{print_full_filter_recode}
\title{Print full filtering for codes for data}
\usage{
print_full_filter(x, conc = TRUE)

print_full_filter_recode(x, conc = TRUE)
}
\arguments{
\item{x}{A data.frame of robonomist id}

\item{conc}{A locigal whether to copy in clipboard}
}
\description{
Can be used also with robonomist id
}
\section{Functions}{
\itemize{
\item \code{print_full_filter_recode}: Code for filter_recode()
}}

\examples{
  print_full_filter(x = data.frame(a = c()), conc = FALSE)
  pttrobo_print_filter_recode(x = "luke/02_Maatalous/06_Talous/02_Maataloustuotteiden_tuottajahinnat/08_Tuottajahinnat_Vilja_rypsi_rapsi_v.px", conc = FALSE)
}
