% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ptt_add_manual_metadata}
\alias{ptt_add_manual_metadata}
\title{Add manual metadata to databases}
\usage{
ptt_add_manual_metadata(db_list_name, x, table_code = NULL)
}
\arguments{
\item{db_list_name}{chr, name of database}

\item{x}{data.frame of manual metadata}

\item{table_code}{chr, code of table}
}
\description{
Add manual metadata to databases
}
