% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ptt_add_pxweb_metadata}
\alias{ptt_add_pxweb_metadata}
\title{Add metadata to databases}
\usage{
ptt_add_pxweb_metadata(db_list_name, table_code = NULL)
}
\arguments{
\item{db_list_name}{chr, name of database}

\item{table_code}{chr, code of table}
}
\description{
Adds pxweb metadata to a db list entry.
}
