% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statfi.R
\name{ptt_capture_citation}
\alias{ptt_capture_citation}
\title{Extract citation information from pxweb_data object.}
\usage{
ptt_capture_citation(x)
}
\arguments{
\item{x, }{pxweb_data object}
}
\value{
list, a list containing citation information
}
\description{
Extract citation information from pxweb_data object.
}
