% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statfi.R
\name{ptt_check_region_classifications}
\alias{ptt_check_region_classifications}
\title{Check if the regions in data correspond to classification keys}
\usage{
ptt_check_region_classifications(data, supress_ok_message = TRUE)
}
\arguments{
\item{supress_ok_message}{logical, whether indicates if no problems. Defaults to TRUE.}
}
\value{
Possibly a warning.
}
\description{
The logic of the function is to take each region level, take each region name-code pair and
check if there is a corresponding region name-code pair in the regional classification key
from the package statficlassifications. Currently the input data has to have the region
variable names 'alue_name' and 'alue_code' for region names and codes respectively.
}
