% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_save_load.R
\name{ptt_create_db_list}
\alias{ptt_create_db_list}
\title{Create full databases}
\usage{
ptt_create_db_list(
  x,
  db_list_name = deparse1(substitute(x)),
  pxweb_metadata = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{a vector of urls or a data.frame with urls as first column and information
on the data corresponding to urls in other columns}

\item{db_list_name}{chr, name of the database}

\item{pxweb_metadata}{logical whether to add pxweb metadata to the database}

\item{overwrite}{logical whether to overwrite possibly existing database}
}
\value{

}
\description{
Create full databases
}
