% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ptt_get_pxweb_metadata}
\alias{ptt_get_pxweb_metadata}
\title{Get pxweb metadata from pxweb api}
\usage{
ptt_get_pxweb_metadata(x, db_list_name = NULL)
}
\arguments{
\item{x}{chr, table_code or url. If table code, function requires db_list_name}

\item{db_list_name}{chr}
}
\value{
pxweb_metadata
}
\description{
Given either a url or a table code and db list name retrieves pxweb_metadata.
}
