% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statfi.R
\name{ptt_get_statfi}
\alias{ptt_get_statfi}
\title{Get PTT data from statfi}
\usage{
ptt_get_statfi(
  url,
  query,
  names = "all",
  renames = NULL,
  check_classifications = TRUE
)
}
\arguments{
\item{url}{A pxweb object or url that can be coherced to a pxweb object}

\item{query}{A json string, json file or list object that can be coherced to a pxweb_query object.}

\item{names}{Whether to add columns for names. "all", "none" or vector of variable names.}
}
\description{
Downloads data from statfi and modifies data to ptt-form.
}
\examples{
  url = "http://pxnet2.stat.fi/PXWeb/api/v1/fi/StatFin/kou/vkour/statfin_vkour_pxt_12bq.px"
  query <-
    list("Vuosi"=c("1970","1975"),
         "Alue"=c("SSS","KU020","KU005"),
         "Ik\U00E4"=c("SSS","15-19"),
         "Sukupuoli"=c("SSS","1","2"),
         "Koulutusaste"=c("SSS","3T8"),
         "Tiedot"=c("vaesto"))

  pp <- ptt_get_statfi(url, query)
  pp <- ptt_get_statfi(url, query, names = "none")
  pp <- ptt_get_statfi(url, query, names = c("Alue"))

  url <- "http://pxnet2.stat.fi/PXWeb/api/v1/fi/Kokeelliset_tilastot/nopsu/koeti_nopsu_pxt_11mx.px"
  query <-
   list("Vuosinelj\U00E4nnes"=c("2019Q1", "2019Q2", "2019Q3", "2019Q4"),
        "Tiedot"=c("alkuperainen_euro",
                   "tyopaivakorjattu_euro",
                   "kausitasoitettu_euro",
                   "trendi_euro"))

  pp2 <- ptt_get_statfi(url, query)
}
