% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_search.R
\name{ptt_glimpse_db}
\alias{ptt_glimpse_db}
\title{Browse metadata attached to a database list}
\usage{
ptt_glimpse_db(db_list_name)
}
\arguments{
\item{db_list_name}{chr, name of db list}
}
\value{

}
\description{
Browse metadata attached to a database list
}
