% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_save_load.R
\name{ptt_save_data}
\alias{ptt_save_data}
\alias{ptt_read_data}
\title{Saves and loads data tables}
\usage{
ptt_save_data(x, x_name = deparse1(substitute(x)), path = NULL)

ptt_read_data(
  x_name,
  region_level = NULL,
  path = NULL,
  only_names = FALSE,
  only_codes = FALSE
)
}
\arguments{
\item{x}{A data to save}

\item{x_name}{A name of data. In saving defaults to a x object name.}

\item{path}{A path to data folder. NULL saves to default PTT-database.}

\item{region_level}{A region level(s) to read.}
}
\description{
For now saves to shared teams
Pellervon taloustutkimus - Datapankki/Tietokanta folder.
The foldet must be synced.
}
\section{Functions}{
\itemize{
\item \code{ptt_read_data}: Read data
}}

\examples{
test_dat <- data.frame(alue_name = factor(c("KOKO MAA", "Uusimaa")),
                       alue_code = factor(c("SSS", "MK01")),
                       a = c(1,2))
ptt_save_data(test_dat)
test_dat2 <- ptt_read_data("test_dat")
}
