% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{ptt_select}
\alias{ptt_select}
\title{Select columns and filter SSS from columns not selected}
\usage{
ptt_select(data, ..., SSS = FALSE)
}
\arguments{
\item{data}{data.frame to modify}

\item{...}{chr, column names to select}

\item{SSS}{logical, whether to leave SSS to selected columns}
}
\value{
data.frame
}
\description{
Only removes columns that contain variable SSS. Note that the data has
to have codes.
}
\examples{
data <- cbind(tidyr::crossing(var1 = c("SSS", 1, 2),
                       var2 = c("SSS", "a", "b"),
                       var3 = c("SSS", "c", "d")),
              value = rnorm(27))
data |> ptt_select(var3, value)
data |> ptt_select(var2, var3, value)

}
