% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_names.R
\name{pxweb_metadata_to_codes_names}
\alias{pxweb_metadata_to_codes_names}
\title{Get code-name key from pxweb_metadata object.}
\usage{
pxweb_metadata_to_codes_names(pxweb_metadata)
}
\arguments{
\item{pxweb_metadata}{pxweb_metadata object}
}
\value{
named vector, code-name key
}
\description{
Get code-name key from pxweb_metadata object.
}
