% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_utils.R
\name{pxweb_print_code_full_query}
\alias{pxweb_print_code_full_query}
\title{Print a code to get a full query from table url}
\usage{
pxweb_print_code_full_query(url, time_all = TRUE)
}
\arguments{
\item{url}{An api url to the table or an url to web page.}

\item{time_all}{if TRUE (default) will print time variable as \code{c("*")}.}
}
\description{
Print a code to get a full query from table url
}
\examples{
  pxweb_print_code_full_query(url = statfi_url("StatFin", "vrm/tyokay/statfin_tyokay_pxt_115u.px"))

}
