% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extra.R
\name{rebase}
\alias{rebase}
\title{Rebase (or base) index}
\usage{
rebase(x, time, baseyear, basevalue = 100)
}
\arguments{
\item{x}{a numeric vector. An index to rebase}

\item{time}{a time variable in a Date format.}

\item{baseyear}{a numeric year or vector of years.}

\item{basevalue}{index base values. if NULL value of x at base year.}
}
\description{
Rebase (or base) index
}
