% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_statfi.R
\name{rm_empty_cols}
\alias{rm_empty_cols}
\title{Remove columns with unique value.}
\usage{
rm_empty_cols(data)
}
\arguments{
\item{data}{a data.frame with columns with unique values}
}
\value{
data.frame
}
\description{
Remove columns with unique value.
}
\examples{
 data <- data.frame(var1 = letters[1:10],
                    var2 = rnorm(10),
                    var3 = "a")

 data <- rm_empty_cols(data)
}
