% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfi_utils.R
\name{statfi_url}
\alias{statfi_url}
\title{Statfi url}
\usage{
statfi_url(..., .base_url = "https://pxnet2.stat.fi/PXWeb/api/v1/fi")
}
\arguments{
\item{...}{Character vectors.}

\item{.base_url}{A base url for statfi.}
}
\description{
Gives full statfi url
}
\examples{
  statfi_url("StatFin", "kou/vkour/statfin_vkour_pxt_12bq.px")

}
