% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{agg_abolished_mun}
\alias{agg_abolished_mun}
\title{Aggregate abolished municipalicities}
\usage{
agg_abolished_mun(x)
}
\arguments{
\item{x}{A data.frame like object.}
}
\description{
Aggregate abolished municipalicities
}
\examples{
x <- tibble(alue_code = c("KU911", "KU541"), alue_name = c("A", "B"), values = c(1,2))

}
