% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{filter_recode}
\alias{filter_recode}
\title{Filtering and recoding data}
\usage{
filter_recode(dat, ..., query = NULL, droplevels = TRUE, check = FALSE)
}
\arguments{
\item{dat}{A data.frame to filter and recode.}

\item{...}{A (named) vector with a name.}

\item{query}{A same as ... as a list. Overrides dots.}

\item{droplevels}{A locigal to droplevels of factors. Default TRUE.}

\item{check}{Whether to check missing levels.}
}
\description{
Filters data based on named vectors or list of them.
Filtered series can be recoded at the same time, and
also columns are returned as factor with levels in
order of filtering information.
}
\details{
Gives message for missing filtering levels.
}
\examples{
data.frame(tieto = c("taso kaksi", "taso kaksi", "taso yksi"), tieto2 = c("a", "b", "c")) |>
  filter_recode(
  tieto = c("Yksi" = "taso yksi",
            "Kaksi" = "taso kaksi"),
  tieto2 = c("b", "c")) |>
  str()

# Puuttuvan tiedon tarkistaminen
data.frame(tieto = c("taso kaksi", "taso kaksi", "taso yksi"), tieto2 = c("a", "b", "c")) |>
  filter_recode(
  tieto = c("Yksi" = "taso yksi",
            "Kaksi" = "taso kaksi",
            "Kolme" = "taso kolme"),
  tieto2 = c("a", "b", "c", "d"), check = TRUE) |>
  str()

data.frame(tieto = c("taso kaksi", "taso kaksi", "taso yksi"), tieto2 = c("a", "b", "c")) |>
  filter_recode(
  tieto = c("Yksi" = "taso"),
  tieto2 = c("a", "b", "c"), check = TRUE) |>
  str()
}
