% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_save_load.R
\name{ptt_create_db_list}
\alias{ptt_create_db_list}
\alias{ptt_create_db_list_vector}
\alias{ptt_create_db_list_df}
\title{Create full database lists}
\usage{
ptt_create_db_list(
  x,
  db_list_name = deparse1(substitute(x)),
  pxweb_metadata = TRUE,
  overwrite = FALSE,
  call = "ptt_get_statfi(url, query)"
)

ptt_create_db_list_vector(
  x,
  db_list_name = deparse1(substitute(x)),
  pxweb_metadata = TRUE,
  call = "ptt_get_statfi(url, query)"
)

ptt_create_db_list_df(
  x,
  db_list_name = deparse1(substitute(x)),
  pxweb_metadata = TRUE,
  call = "ptt_get_statfi(url, query)"
)
}
\arguments{
\item{x}{a vector of urls or a data.frame with urls as first column and information
on the data corresponding to urls in other columns}

\item{db_list_name}{chr, name of the database}

\item{pxweb_metadata}{logical whether to add pxweb metadata to the database}

\item{overwrite}{logical whether to overwrite possibly existing database}
}
\description{
Create full database lists
}
\section{Functions}{
\itemize{
\item \code{ptt_create_db_list_vector()}: Create full database lists

\item \code{ptt_create_db_list_df()}: Create full database lists

}}
