% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_save_load.R
\name{ptt_save_data}
\alias{ptt_save_data}
\alias{ptt_read_data}
\title{Saves and loads data tables}
\usage{
ptt_save_data(x, x_name = deparse1(substitute(x)))

ptt_read_data(x_name, region_level = NULL)
}
\arguments{
\item{x}{A data to save}

\item{x_name}{A name of data. In saving defaults to a x object name.}
}
\description{
For now saves to shared teams
Pellervon taloustutkimus - Datapankki/Tietokanta folder.
The foldet must be synced.
}
\section{Functions}{
\itemize{
\item \code{ptt_read_data}: Read data
}}

\examples{
test_dat <- data.frame(a = c(1,2))
ptt_save_data(test_dat)
test_dat2 <- ptt_read_data("test_dat")
}
