% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptt_search.R
\name{ptt_search}
\alias{ptt_search}
\alias{ptt_search_data}
\alias{ptt_search_db}
\title{Search PTT data bases}
\usage{
ptt_search(..., filetype = NULL, path = db_path)

ptt_search_data(..., path = db_path)

ptt_search_db(..., path = db_path)
}
\arguments{
\item{...}{a search term}

\item{filetype}{"qs" or "rds"}
}
\value{

}
\description{
Search PTT data bases
}
\section{Functions}{
\itemize{
\item \code{ptt_search_data}: Search data sets in PTT database

\item \code{ptt_search_db}: Search database lists in PTT database
}}

