% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb_utils.R
\name{pxweb_print_full_query}
\alias{pxweb_print_full_query}
\title{Print full query from table url}
\usage{
pxweb_print_full_query(url)
}
\arguments{
\item{url}{A url to the table}
}
\description{
Print full query from table url
}
\examples{
  pxweb_print_full_query(url = statfi_url("StatFin", "vrm/tyokay/statfin_tyokay_pxt_115u.px"))

}
