% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfi_utils.R
\name{statfi_parse_url}
\alias{statfi_parse_url}
\title{Parse Statfi pxweb url from web url}
\usage{
statfi_parse_url(url)
}
\arguments{
\item{url}{An url from web to parse}
}
\description{
Parse Statfi pxweb url from web url
}
\examples{
  statfi_parse_url("https://pxnet2.stat.fi/PXWeb/pxweb/fi/StatFin/StatFin__vrm__muutl/statfin_muutl_pxt_119z.px/")

}
